+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- Log/trace JSON's size in the Assets client's `getAppJSONByVendor` method.
+
 ## [6.45.22] - 2023-09-20
 
 ### Removed